pyproject.toml
[tool.poetry]
name = "websitetemplate-mark-03"
version = "0.1.0"
description = ""
authors = ["Vagner Bessa"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.12.12"
jinja2 = "^3.1.6"
python-dotenv = "^1.1.0"
aiohttp-jinja2 = "^1.6"


[tool.poetry.group.dev.dependencies]
watchfiles = "^1.0.5"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.2.0"
black = "^25.1.0"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
